Overhauled structure, authentication, authorization, CartItem service, comments, many other changes

+++ OctopusStore/ClientApp/src/app/item-image/item-image-create-update/item-image-create-update.component.ts
@@ -1,9 +1,11 @@
-import { Component, OnInit, Input } from '@angular/core';
+import { Component, OnInit, Input, ViewChild } from '@angular/core';
 import { ItemImage } from '../../view-models/item-image/item-image';
 import { ActivatedRoute } from '@angular/router';
 import { ItemImageService } from '../../services/item-image-service';
 import { ItemImageIndex } from '../../view-models/item-image/item-image-index';
 import { Item } from '../../view-models/item/item';
+import { Subscription } from 'rxjs';
+import { MessageService } from '../../services/message.service';
 
 @Component({
   selector: 'app-item-image-create-update',
@@ -11,14 +13,16 @@ import { Item } from '../../view-models/item/item';
   styleUrls: ['./item-image-create-update.component.css']
 })
 export class ItemImageCreateUpdateComponent implements OnInit {
-
+  itemImageIndexIndexSubscription: Subscription;
+  itemImageSubscription: Subscription;
   itemImageFormData: FormData;
   itemImages: ItemImage[];
   @Input() item: Item;
 
+
   constructor(
     private itemImageService: ItemImageService,
-    private route: ActivatedRoute) {
+    private messageService: MessageService) {
     this.itemImages = [];
   }
 
@@ -28,38 +32,46 @@ export class ItemImageCreateUpdateComponent implements OnInit {
 
   initializeComponent() {
     if (this.item.id) {
-      this.itemImageService.index({ itemId: this.item.id }).subscribe((data: ItemImageIndex) => {
-        this.itemImages.push.apply(this.itemImages, data.entities);
-      })
+      this.itemImageService.index({ itemId: this.item.id }).subscribe((itemImageIndex: ItemImageIndex) => {
+        itemImageIndex.entities.forEach(i => this.itemImages.push(new ItemImage(i)));
+      });
     }
   }
   saveItemImage() {
-    this.itemImageService.postFormData(this.itemImageFormData, this.itemImageService.filePostUrl).subscribe((data: ItemImage) => {
-      if (data)
-        this.itemImages.push(data);
+    this.itemImageService.postFile(this.itemImageFormData, this.item.id).subscribe((data: ItemImage) => {
+      if (data) {
+        this.itemImages.push(new ItemImage(data));
+        this.messageService.sendSuccess("Item image saved");
+      }
     });
   }
   updateItemImage(itemImage: ItemImage) {
     let i = this.itemImages.indexOf(itemImage);
-    this.itemImageService.update(itemImage).subscribe((data: ItemImage) => {
-      if (data)
-        this.itemImages[i] = data;
-    });
+    this.itemImageService.postOrPut(itemImage).subscribe((data: ItemImage) => {
+      if (data) {
+        this.itemImages[i] = new ItemImage(data);
+        this.messageService.sendSuccess("Item image updated");
+      }
+    });;
   }
   getImageUrl(itemImage: ItemImage): string {
     return this.itemImageService.getImageUrl(itemImage.id)
   }
 
   deleteItemImage(itemImage: ItemImage) {
-    this.itemImageService.delete(itemImage.id).subscribe((data) => {
-      if (data)
-        this.itemImages = this.itemImages.filter(i => i != itemImage);
-    })
+    if (itemImage.id) {
+      this.itemImageService.delete(itemImage.id).subscribe((data) => {
+        if (data) {
+          this.itemImages = this.itemImages.filter(i => i != itemImage);
+          this.messageService.sendSuccess("Item image deleted");
+        }
+      })
+    }
   }
 
   onFileChanged(event) {
     let selectedFile = event.target.files[0];
     this.itemImageFormData = new FormData();
-    this.itemImageFormData.append('formFile', selectedFile, `${this.item.id}`);
+    this.itemImageFormData.append('formFile', selectedFile, selectedFile.name);
   }
 }
